Add files via upload
Some changes were made in structure and coordinates to discover why sprites weren't loading. Upon discovery, I realized it's because the x and y coordinates are labeled as undefined and I haven't discerned why that is. My code follows similar structure as Marriott's with changes made necessary to accommodate for additional animations.